+++ click_extended/decorators/convert/__init__.py
@@ -1,5 +1,8 @@
 """Initialization file for the `click_extended.decorators.convert` module."""
 
+from click_extended.decorators.convert.convert_byte_size import (
+    convert_byte_size,
+)
 from click_extended.decorators.convert.convert_temperature import (
     convert_temperature,
 )
@@ -7,6 +10,7 @@ from click_extended.decorators.convert.convert_time import convert_time
 from click_extended.decorators.convert.convert_weight import convert_weight
 
 __all__ = [
+    "convert_byte_size",
     "convert_time",
     "convert_temperature",
     "convert_weight",